fix(protected_routes): - Applied protected routes for User Page, Explore Page, and Propose Page. - Implemented name & email retrieval from registered users.
