+++ src/app/Components/ProtectedRoute.tsx
@@ -1,23 +1,33 @@
 "use client";
 
 import { useEffect, useState } from "react";
-import { useRouter } from "next/navigation";
+import { useRouter, usePathname } from "next/navigation";
 import { getCurrentUser } from "../services/auth";
 
 interface ProtectedRouteProps {
   children: React.ReactNode | ((user: any) => React.ReactNode);
-  redirectTo?: string; 
+  redirectTo?: string;
 }
 
-const ProtectedRoute = ({ children, redirectTo = "/login" }: ProtectedRouteProps) => {
+const ProtectedRoute = ({
+  children,
+  redirectTo = "/login",
+}: ProtectedRouteProps) => {
   const [user, setUser] = useState<any>(null);
   const [loading, setLoading] = useState(true);
   const router = useRouter();
+  const pathname = usePathname(); // Get current route
 
   useEffect(() => {
     const checkAuth = async () => {
       const userData = await getCurrentUser();
-      if (!userData) {
+
+      // If accessing /user-profile and not logged in, redirect to /register
+      if (!userData && pathname === "/user-profile") {
+        router.push("/login");
+      }
+      // Redirect other protected routes to login
+      else if (!userData) {
         router.push(redirectTo);
       } else {
         setUser(userData);
@@ -26,13 +36,17 @@ const ProtectedRoute = ({ children, redirectTo = "/login" }: ProtectedRouteProps
     };
 
     checkAuth();
-  }, [router, redirectTo]);
+  }, [router, redirectTo, pathname]);
 
-  if (loading) return <p>Loading...</p>; 
+  if (loading) return <p>Loading...</p>;
 
   if (!user) return null; // Prevents UI flickering before redirect
 
-  return typeof children === "function" ? <>{children(user)}</> : <>{children}</>;
+  return typeof children === "function" ? (
+    <>{children(user)}</>
+  ) : (
+    <>{children}</>
+  );
 };
 
 export default ProtectedRoute;
