+++ src/app/components/Navbar.tsx
@@ -40,7 +40,9 @@ const Navbar = () => {
     <nav className="fixed top-0 left-0 w-full flex items-center justify-between px-12 py-4 bg-white dark:bg-gray-900 shadow-md z-50 h-16">
       <div className="flex items-center">
         <Image
-          src={darkMode ? "/assets/Lightlogo.png" : "/assets/Dark_logo_projc_1.png"}
+          src={
+            darkMode ? "/assets/Lightlogo.png" : "/assets/Dark_logo_projc_1.png"
+          }
           alt="Proj.C Logo"
           width={160}
           height={70}
@@ -50,19 +52,31 @@ const Navbar = () => {
 
       {pathname === "/home" && (
         <div className="hidden md:flex flex-1 justify-center space-x-10 text-gray-900 dark:text-white font-medium">
-          <Link href="/home" className="relative font-semibold hover:text-purple-600 dark:hover:text-purple-400 transition-all">
+          <Link
+            href="/home"
+            className="relative font-semibold hover:text-purple-600 dark:hover:text-purple-400 transition-all"
+          >
             Home
             {pathname === "/home" && (
               <span className="absolute left-0 bottom-[-4px] w-full h-[2px] bg-purple-600 dark:bg-purple-400"></span>
             )}
           </Link>
-          <Link href="#" className="hover:text-purple-600 dark:hover:text-purple-400 transition-all font-semibold">
+          <Link
+            href="#"
+            className="hover:text-purple-600 dark:hover:text-purple-400 transition-all font-semibold"
+          >
             Notifications
           </Link>
-          <Link href="#" className="hover:text-purple-600 dark:hover:text-purple-400 transition-all font-semibold">
+          <Link
+            href="/project-explore"
+            className="hover:text-purple-600 dark:hover:text-purple-400 transition-all font-semibold"
+          >
             Explore
           </Link>
-          <Link href="#" className="hover:text-purple-600 dark:hover:text-purple-400 transition-all font-semibold">
+          <Link
+            href="/user-profile"
+            className="hover:text-purple-600 dark:hover:text-purple-400 transition-all font-semibold"
+          >
             Profile
           </Link>
         </div>
@@ -71,9 +85,15 @@ const Navbar = () => {
       <div className="flex items-center space-x-6">
         {/* Toggle between sun and moon icon */}
         {darkMode ? (
-          <IoSunnyOutline onClick={toggleDarkMode} className="text-2xl cursor-pointer text-yellow-400 hover:text-yellow-500 transition-all" />
+          <IoSunnyOutline
+            onClick={toggleDarkMode}
+            className="text-2xl cursor-pointer text-yellow-400 hover:text-yellow-500 transition-all"
+          />
         ) : (
-          <IoMoonOutline onClick={toggleDarkMode} className="text-2xl cursor-pointer text-gray-900 dark:text-white hover:text-purple-600 dark:hover:text-purple-400 transition-all" />
+          <IoMoonOutline
+            onClick={toggleDarkMode}
+            className="text-2xl cursor-pointer text-gray-900 dark:text-white hover:text-purple-600 dark:hover:text-purple-400 transition-all"
+          />
         )}
 
         <button
