--- src/app/home-page/page.tsx
@@ -1,15 +1,0 @@
-import React from 'react'
-import Navbar from '../Components/Navbar/page'
-import Footer from '../Components/Footer'
-
-const page = () => {
-  return (
-    <>
-    <Navbar/>
-    <h3>Hi guys this is homepage</h3>
-    <Footer/>
-    </>
-  )
-}
-
-export default page
